refactoring UserServices.js helper and auth route
